# src/app/assets/swagger-manager.json
{
	"openapi":"3.1.0",
	"info":{
		"title":"Rest API Manager",
		"description":"{{SWAGGER_INFO_DESCRIPTION_MNG}}",
		"version":"v0.1.0",
		"contact":{
			"name":"API Support",
			"url":"https://github.com/tglima/restapi-node-docker"
		},
		"license":{
			"name":"MIT",
			"url":"https://raw.githubusercontent.com/tglima/restapi-node-docker/main/LICENSE"
		}
	},
	"servers":[
		{
			"url":"/v{{NU_VERSION}}/mng",
			"description":"{{SWAGGER_SERVERS_DESCRIPTION}}"
		}
	],
	"paths":{
...
		"/log-errors/find?id_log_error={id_log_error}":{
			"get":{
				"tags":[
					"LogError"
				],
				"summary":"Retorna LogError através do id informado",
				"operationId":"findLogErrorById",
				"security":[
					{
						"API_KEY":[

						],
						"AUTHORIZATION":[

						]
					}
				],
				"parameters":[
					{
						"name":"id_log_error",
						"in":"query",
						"description":"Id único de cada LogError",
						"required":true,
						"schema":{
							"type":"integer",
							"format":"int32"
						}
					}
				],
				"responses":{
					"200":{
						"description":"LogError encontrado",
						"content":{
							"application/json":{
								"schema":{
									"$ref":"#/components/schemas/LogError"
								}
							}
						}
					},
					"400":{
						"description":"Falha na requisição",
						"content":{
							"application/json":{
								"schema":{
									"$ref":"#/components/schemas/DefaultResponse"
								},
								"examples":{
									"InvalidRequestResponse":{
										"$ref":"#/components/examples/InvalidRequestResponse"
									}
								}
							}
						}
					},
					"401":{
						"description":"Falha na autenticação",
						"content":{
							"application/json":{
								"schema":{
									"$ref":"#/components/schemas/DefaultResponse"
								},
								"examples":{
									"InvalidRequestResponse":{
										"$ref":"#/components/examples/AuthFailedResponse"
									}
								}
							}
						}
					},
					"404":{
						"description":"LogError não encontrado",
						"content":{
							"application/json":{
								"schema":{
									"$ref":"#/components/schemas/DefaultResponse"
								},
								"examples":{
									"NotFoundResponse":{
										"$ref":"#/components/examples/NotFoundResponse"
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/database-info":{
			"get":{
				"tags":[
					"Database"
				],
				"summary":"Retorna informações relevantes referentes ao banco de dados (Tamanho em mb, nome das tabelas, total de registros, etc)",
				"operationId":"GetDatabaseInfo",
				"security":[
					{
						"API_KEY":[

						],
						"AUTHORIZATION":[

						]
					}
				],
				"parameters":[

				],
				"responses":{
					"200":{
						"description":"Dados relevantes do banco de dados",
						"content":{
							"application/json":{
								"schema":{
									"$ref":"#/components/schemas/DatabaseInfo"
								},
								"examples":{
									"DatabaseInfo":{
										"$ref":"#/components/examples/DatabaseInfo"
									}
								}
							}
						}
					},
					"400":{
						"description":"Falha na requisição",
						"content":{
							"application/json":{
								"schema":{
									"$ref":"#/components/schemas/DefaultResponse"
								}
							}
						}
					},
					"401":{
						"description":"Falha na autenticação",
						"content":{
							"application/json":{
								"schema":{
									"$ref":"#/components/schemas/DefaultResponse"
								},
								"examples":{
									"InvalidRequestResponse":{
										"$ref":"#/components/examples/AuthFailedResponse"
									}
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components":{
		"schemas":{
...
			"LogError":{
				"type":"object",
				"properties":{
					"id_log_error":{
						"type":"integer"
					},
					"dt_register":{
						"type":"string",
						"format":"date-time"
					},
					"method":{
						"type":"string"
					},
					"json_log_error":{
						"type":"object"
					}
				}
			},
			"DatabaseInfo":{
				"type":"object",
				"properties":{
					"database_name":{
						"type":"string"
					},
					"database_size":{
						"type":"string"
					},
					"tables_name":{
						"type":"array",
						"items":{
							"type":"string"
						}
					},
					"product_total":{
						"type":"integer"
					},
					"log_event_total":{
						"type":"integer"
					},
					"log_error_total":{
						"type":"integer"
					}
				}
			},
...
			"DefaultResponse":{
				"type":"object",
				"properties":{
					"code_event":{
						"type":"string"
					},
					"messages":{
						"type":"array",
						"items":{
							"type":"string"
						}
					}
				}
			},
...
		},
		"examples":{
			"AuthFailedResponse":{
				"value":{
					"code_event":"b000b00a-ab00-0000-b000-bb000b000000",
					"messages":[
						"Credenciais inválidas ou ausentes"
					]
				}
			},
			"InvalidRequestResponse":{
				"value":{
					"code_event":"b000b00a-ab00-0000-b000-bb000b000000",
					"messages":[
						"Ocorreu uma falha/erro na sua requisição. Reveja os dados enviados e tente novamente!"
					]
				}
			},
			"NotFoundResponse":{
				"value":{
					"code_event":"b000b00a-ab00-0000-b000-bb000b000000",
					"messages":[
						"Item não encontrado!"
					]
				}
			},
			"DatabaseInfo":{
				"value":{
					"database_name":"database.db",
					"database_size":"1.5 mb",
					"tables_name":[
						"product",
						"log_event",
						"log_error"
					],
					"product_total":3,
					"log_event_total":30,
					"log_error_total":0
				}
			},
...
		},
		"securitySchemes":{
			"API_KEY":{
				"type":"apiKey",
				"in":"header",
				"name":"API-KEY",
				"description":"Chave de acesso individual disponibilizado para acessar a API"
			},
			"AUTHORIZATION":{
				"type":"apiKey",
				"in":"header",
				"name":"authorization",
				"description":"Chave de acesso disponibilizada apenas para admnistradores da API"
			}
		}
	}
}